Poison Breath and Mortar

+++ Bouncy/js/Abilities/AbilityCores/NewCores/CoreCreator.py
@@ -2,6 +2,7 @@ import fileinput
 import argparse
 
 from shutil import copyfile
+from pathlib import Path
 
 parser = argparse.ArgumentParser(description='Make a bunch of ability cores.')
 parser.add_argument('coreNum', metavar='N', type=int,
@@ -10,17 +11,24 @@ parser.add_argument('coreNum', metavar='N', type=int,
 args = parser.parse_args()
 
 coreNum = args.coreNum
-if coreNum:
-    fileName = "AbilityCore" + str(coreNum) + ".js"
+try:
+    if coreNum:
+        fileName = "AbilityCore" + str(coreNum) + ".js"
 
-    inputFile = open("AbilityCoreTemplate.js", 'r')
-    outputFile = open(fileName, 'w')
-    #fileToSearch = 'D:\dummy1.txt'
+        my_file = Path(fileName)
+        if my_file.exists():
+            raise Exception("This ability already exists");
 
-    textToSearch = "NUMBER"
-    fileContents = inputFile.read()
-    inputFile.close()
+        inputFile = open("AbilityCoreTemplate.js", 'r')
+        outputFile = open(fileName, 'w')
+        #fileToSearch = 'D:\dummy1.txt'
 
-    fileContents = fileContents.replace("NUMBER", str(coreNum))
-    outputFile.write(fileContents)
-    outputFile.close()
+        textToSearch = "NUMBER"
+        fileContents = inputFile.read()
+        inputFile.close()
+
+        fileContents = fileContents.replace("NUMBER", str(coreNum))
+        outputFile.write(fileContents)
+        outputFile.close()
+except Exception as error:
+    print ("Error: " + repr(error));
